rename fil_fungible_token to frc46_token
