--- fil_fungible_token/README.md
@@ -1,9 +1,0 @@
-# fil_fungible_token
-
-This library handles the state-level accounting and logic for implementing an
-FRCXXX token.
-
-## WIP
-
-Discussions around API and functionality can be found
-[here](https://github.com/filecoin-project/FIPs/discussions/407)
