+++ frc46_token/Cargo.toml
@@ -1,12 +1,16 @@
 [package]
-name = "fil_fungible_token"
+name = "frc46_token"
+description = "Filecoin FRC-0046 fungible token reference implementation"
 version = "0.1.0"
+license = "MIT OR Apache-2.0"
+keywords = ["filecoin", "fvm", "token", "frc-0046"]
+repository = "https://github.com/helix-onchain/filecoin/"
 edition = "2021"
 
 [dependencies]
 anyhow = "1.0.56"
 cid = { version = "0.8.3", default-features = false, features = ["serde-codec"] }
-frc42_dispatch = { path = "../frc42_dispatch" }
+frc42_dispatch = { version = "0.3.0", path = "../frc42_dispatch" }
 fvm_ipld_blockstore = "0.1.1"
 fvm_ipld_hamt = "0.5.1"
 fvm_ipld_amt = { version = "0.4.2", features = ["go-interop"] }
