+++ frc46_token/README.md
@@ -1,0 +1,6 @@
+# frc46_token
+
+This is the reference Rust library that implements the
+[FRC0046 fungible token standard](https://github.com/filecoin-project/FIPs/blob/master/FRCs/frc-0046.md).
+It is intended for use in native user-programmable actors deployed to the
+Filecoin Virtual Machine.
