+++ testing/fil_token_integration/Cargo.toml
@@ -9,7 +9,7 @@ anyhow = { version = "1.0.63", features = ["backtrace"] }
 cid = { version = "0.8.5", default-features = false }
 fvm = { version = "2.0.0-alpha.2", default-features = false }
 frc42_dispatch = { path = "../../frc42_dispatch" }
-fil_fungible_token = { version = "0.1.0", path = "../../fil_fungible_token" }
+frc46_token = { version = "0.1.0", path = "../../frc46_token" }
 fvm_integration_tests = "2.0.0-alpha.1"
 fvm_ipld_blockstore = "0.1.1"
 fvm_ipld_encoding = "0.2.2"
