+++ testing/fil_token_integration/actors/basic_nft_actor/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-fil_fungible_token = { path = "../../../../fil_fungible_token" }
+frc46_token = { path = "../../../../frc46_token" }
 frcxx_nft = { path = "../../../../frcxx_nft" }
 frc42_dispatch = { path = "../../../../frc42_dispatch" }
 fvm_ipld_blockstore = "0.1.1"
