+++ testing/fil_token_integration/actors/basic_receiving_actor/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-fil_fungible_token = { version = "0.1.0", path = "../../../../fil_fungible_token" }
+frc46_token = { version = "0.1.0", path = "../../../../frc46_token" }
 frc42_dispatch = { path = "../../../../frc42_dispatch" }
 fvm_ipld_blockstore = "0.1.1"
 fvm_ipld_encoding = "0.2.2"
