+++ testing/fil_token_integration/actors/basic_token_actor/Cargo.toml
@@ -9,7 +9,7 @@ fvm_ipld_blockstore = { version = "0.1.1" }
 fvm_ipld_encoding = { version = "0.2.2" }
 fvm_sdk = { version = "2.0.0-alpha.2" }
 fvm_shared = { version = "2.0.0-alpha.2" }
-fil_fungible_token = { version = "0.1.0", path = "../../../../fil_fungible_token" }
+frc46_token = { version = "0.1.0", path = "../../../../frc46_token" }
 num-traits = { version = "0.2.15" }
 serde = { version = "1.0.136", features = ["derive"] }
 serde_tuple = { version = "0.5.0" }
