+++ testing/fil_token_integration/actors/basic_transfer_actor/Cargo.toml
@@ -5,7 +5,7 @@ edition = "2021"
 
 [dependencies]
 cid = { version = "0.8.5", default-features = false }
-fil_fungible_token = { version = "0.1.0", path = "../../../../fil_fungible_token" }
+frc46_token = { version = "0.1.0", path = "../../../../frc46_token" }
 frc42_dispatch = { path = "../../../../frc42_dispatch" }
 fvm_ipld_blockstore = { version = "0.1.1" }
 fvm_ipld_encoding = { version = "0.2.2" }
