+++ testing/fil_token_integration/actors/basic_transfer_actor/src/lib.rs
@@ -1,9 +1,9 @@
 use cid::{multihash::Code, Cid};
-use fil_fungible_token::{
+use frc42_dispatch::{match_method, method_hash};
+use frc46_token::{
     receiver::types::{FRC46TokenReceived, UniversalReceiverParams, FRC46_TOKEN_TYPE},
     token::types::TransferParams,
 };
-use frc42_dispatch::{match_method, method_hash};
 use fvm_ipld_blockstore::Block;
 use fvm_ipld_encoding::tuple::{Deserialize_tuple, Serialize_tuple};
 use fvm_ipld_encoding::{de::DeserializeOwned, RawBytes, DAG_CBOR};
