+++ testing/fil_token_integration/tests/mint_fungible_tokens.rs
@@ -2,8 +2,8 @@ use std::env;
 
 use basic_token_actor::MintParams;
 use cid::Cid;
-use fil_fungible_token::token::{state::TokenState, types::MintReturn};
 use frc42_dispatch::method_hash;
+use frc46_token::token::{state::TokenState, types::MintReturn};
 use fvm::executor::{ApplyKind, Executor};
 use fvm_integration_tests::bundle;
 use fvm_integration_tests::dummy::DummyExterns;
